Move forch entry point to __main__.py
