+++ forch/forchestrator.py
@@ -1,11 +1,8 @@
 """Orchestrator component for controlling a Faucet SDN"""
 
 from datetime import datetime
-import argparse
-import functools
 import logging
 import os
-import sys
 import threading
 import time
 import yaml
@@ -37,12 +34,16 @@ from forch.proto.system_state_pb2 import SystemState
 
 LOGGER = logging.getLogger('forch')
 
-_FORCH_CONFIG_DEFAULT = 'forch.yaml'
 _STRUCTURAL_CONFIG_DEFAULT = 'faucet.yaml'
 _BEHAVIORAL_CONFIG_DEFAULT = 'faucet.yaml'
 _SEGMENTS_VLAN_DEFAULT = 'segments-to-vlans.yaml'
 _DEFAULT_PORT = 9019
 _PROMETHEUS_HOST = '127.0.0.1'
+
+
+# TODO: 1) use protobuf after entire forch config is converted
+#       2) clean up places where individual forch config sections are converted by dict_proto
+#          instead of direct access from forch config proto obj
 
 class Forchestrator:
     """Main class encompassing faucet orchestrator components for dynamically
@@ -595,85 +596,3 @@ class Forchestrator:
             return self._augment_state_reply(reply, path)
         except Exception as e:
             return f"Cannot read faucet config: {e}"
-
-
-def load_config():
-    """Load configuration from the configuration file"""
-    # TODO: 1) use protobuf after entire forch config is converted
-    #       2) clean up places where individual forch config sections are converted by dict_proto
-    #          instead of direct access from forch config proto obj
-    config_root = os.getenv('FORCH_CONFIG_DIR', '.')
-    config_path = os.path.join(config_root, _FORCH_CONFIG_DEFAULT)
-    LOGGER.info('Reading config file %s', os.path.abspath(config_path))
-    try:
-        with open(config_path, 'r') as stream:
-            return yaml.safe_load(stream)
-    except Exception as e:
-        LOGGER.error('Cannot load config: %s', e)
-        return None
-
-
-def show_error(error, path, params):
-    """Display errors"""
-    return f"Cannot initialize forch: {str(error)}"
-
-
-def main():
-    """main function to start forch"""
-    configure_logging()
-
-    config = load_config()
-    if not config:
-        LOGGER.error('Invalid config, exiting.')
-        sys.exit(1)
-
-    forchestrator = Forchestrator(config)
-    http_server = forch.http_server.HttpServer(
-        config.get('http', {}), forchestrator.get_local_port())
-
-    try:
-        forchestrator.initialize()
-        http_server.map_request('system_state', forchestrator.get_system_state)
-        http_server.map_request('dataplane_state', forchestrator.get_dataplane_state)
-        http_server.map_request('switch_state', forchestrator.get_switch_state)
-        http_server.map_request('cpn_state', forchestrator.get_cpn_state)
-        http_server.map_request('process_state', forchestrator.get_process_state)
-        http_server.map_request('host_path', forchestrator.get_host_path)
-        http_server.map_request('list_hosts', forchestrator.get_list_hosts)
-        http_server.map_request('sys_config', forchestrator.get_sys_config)
-        http_server.map_request('', http_server.static_file(''))
-    except Exception as e:
-        LOGGER.error("Cannot initialize forch: %s", e, exc_info=True)
-        http_server.map_request('', functools.partial(show_error, e))
-    finally:
-        http_server.start_server()
-
-    if forchestrator.initialized():
-        forchestrator.main_loop()
-    else:
-        try:
-            http_server.join_thread()
-        except KeyboardInterrupt:
-            LOGGER.info('Keyboard interrupt. Exiting.')
-
-    LOGGER.warning('Exiting program')
-    http_server.stop_server()
-    forchestrator.stop()
-
-
-def parse_args(raw_args):
-    """Parse command line arguments"""
-    parser = argparse.ArgumentParser(prog='forch', description='Process some integers.')
-    parser.add_argument('-V', '--version', action='store_true', help='print version and exit')
-    parsed = parser.parse_args(raw_args)
-    return parsed
-
-
-if __name__ == '__main__':
-    ARGS = parse_args(sys.argv[1:])
-
-    if ARGS.version:
-        print(f'Forch version {__version__}')
-        sys.exit()
-
-    main()
