+++ setup.cfg
@@ -12,9 +12,14 @@ classifier =
     Natural Language :: English
     Programming Language :: Python :: 3.7
     Operating System :: Unix
+
 [files]
 packages =
     forch
+
+[entry_points]
+console_scripts =
+    forch = forch.__main__:main
 
 [pytype]
 python_version = 3.6
