+++ setup.py
@@ -49,6 +49,5 @@ setuptools.setup(
     ],
     setup_requires=['pbr>=1.9', 'setuptools>=17.1'],
     data_files=data_files,
-    scripts=['bin/forch'],
     pbr=True
 )
